+++ src/main/java/com/example/travel_project/domain/plan/data/Plan.java
@@ -1,0 +1,50 @@
+package com.example.travel_project.domain.plan.data;
+
+import com.example.travel_project.domain.plan.data.mapping.UserPlanList;
+import jakarta.persistence.*;
+import lombok.*;
+import org.springframework.format.annotation.DateTimeFormat;
+
+import java.time.LocalDateTime;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.UUID;
+
+@Entity
+@Getter
+@Setter
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class Plan {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private long id;
+
+    @Column(unique = true, nullable = false, updatable = false)
+    private String uuid;
+
+    private String title;
+
+    @DateTimeFormat(iso = DateTimeFormat.ISO.DATE)
+    private LocalDateTime startDate;
+
+    @DateTimeFormat(iso = DateTimeFormat.ISO.DATE)
+    private LocalDateTime endDate;
+
+    @Lob      // 문자열 길이 길어서 text타입으로 저장
+    private String content;
+
+    @Column(name = "author_email")
+    private String authorEmail;
+
+    @OneToMany(mappedBy = "plan", cascade = CascadeType.ALL, orphanRemoval = true)
+    private List<UserPlanList> userPlanLists = new ArrayList<>();
+
+    @PrePersist
+    public void setUuid() {
+        if (this.uuid == null) {
+            this.uuid = UUID.randomUUID().toString();
+        }
+    }
+}
