+++ src/main/java/com/example/travel_project/domain/plan/web/dto/InviteListResponseDTO.java
@@ -1,0 +1,17 @@
+package com.example.travel_project.domain.plan.web.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class InviteListResponseDTO {
+    private Long id;
+    private String userName;
+    private String userEmail;
+    private String planTitle;
+    private Boolean isAccepted;
+}
+
