+++ src/main/java/com/example/travel_project/domain/plan/web/dto/InviteRequestDTO.java
@@ -1,0 +1,15 @@
+package com.example.travel_project.domain.plan.web.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor
+public class InviteRequestDTO {
+    private String email;
+}
+
