+++ src/main/java/com/example/travel_project/domain/user/repository/UserRepository.java
@@ -1,0 +1,13 @@
+package com.example.travel_project.domain.user.repository;
+
+import com.example.travel_project.domain.user.data.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.Optional;
+
+@Repository
+public interface UserRepository extends JpaRepository<User, Long> {
+    boolean existsByEmail(String email);
+    Optional<User> findByEmail(String email);
+}
