--- src/main/java/com/example/travel_project/dto/PlaceListsDTO.java
@@ -1,16 +1,0 @@
-package com.example.travel_project.dto;
-
-import lombok.*;
-
-import java.util.List;
-
-@Data
-@Getter
-@Builder
-@AllArgsConstructor
-public class PlaceListsDTO {
-    private List<PlaceDTO> attractionList;
-    private List<PlaceDTO> restaurantList;
-    private List<PlaceDTO> cafeList;
-    private List<PlaceDTO> hotelList;
-}
