--- src/main/java/com/example/travel_project/entity/InviteList.java
@@ -1,30 +1,0 @@
-package com.example.travel_project.entity;
-
-import jakarta.persistence.*;
-import lombok.*;
-
-@Entity
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class InviteList {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    // 초대받은 사용자
-    @ManyToOne(fetch = FetchType.LAZY, optional = false)
-    @JoinColumn(name = "user_id", nullable = false)
-    private User user;
-
-    // 초대된 플랜
-    @ManyToOne(fetch = FetchType.LAZY, optional = false)
-    @JoinColumn(name = "plan_id", nullable = false)
-    private Plan plan;
-
-    @Column(name = "is_accepted", nullable = false)
-    private Boolean isAccepted;
-}
