--- src/main/java/com/example/travel_project/repository/UserRepository.java
@@ -1,13 +1,0 @@
-package com.example.travel_project.repository;
-
-import com.example.travel_project.entity.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-
-@Repository
-public interface UserRepository extends JpaRepository<User, Long> {
-    boolean existsByEmail(String email);
-    Optional<User> findByEmail(String email);
-}
